requirements.txt
--extra-index-url=http://dist.pinaxproject.com/dev/
--extra-index-url=http://dist.pinaxproject.com/alpha/

Django==1.4.2
pinax-theme-bootstrap==2.2.1
pinax-theme-bootstrap-account==1.0b2
django-user-accounts==1.0b7
django-forms-bootstrap==2.0.3.post1
metron==1.0
pinax-utils==1.0b1.dev3
django-debug-toolbar==0.8.5
pytz
django-openid
python-openid
django-compressor
django-announcements
django-userena

pip
virtualenv
fabric
gunicorn
psycopg2
PPyGIS
Pillow
South
django-stdimage
sorl-thumbnail
django-urli18n
django-haystack
xapian-haystack
django-hvad
django-olwidget
django-floppyforms
django-rosetta
gevent
eventlet
